bootstrap_css

+++ frontend/app/src/Category.tsx
@@ -2,18 +2,22 @@ import React, { useState, useContext } from 'react';
 import { Link } from 'react-router-dom';
 import { useForm } from 'react-hook-form';
 
+import Button from 'react-bootstrap/Button';
+import Modal from 'react-bootstrap/Modal';
+import Badge from 'react-bootstrap/Badge';
+
 import { useDeleteCategoryMutation } from './hooks/useDeleteCategoryMutation';
 import { useUpdateCategoryMutation } from './hooks/useUpdateCategoryMutation';
 import { CategoryAttributes } from './types/category'
 import { AdminContext } from './utils/AdminProvider';
-import TextField from './uikit/textField';
+import TextField from './uikit/TextField';
 import SubmitButton from './uikit/SubmitButton';
-import FullMessages from './FullMessages';
 
 type Props = {
   category: {
     id: string;
     name: string;
+    postsCnt: number;
   }
 }
 
@@ -24,17 +28,19 @@ type CategoryInputs = {
 
 export default function Category(props: Props) {
   const currentAdmin = useContext(AdminContext);
-  const { category: { id, name } } = props;
-  const [fullMessages, setFullMessages] = useState([]);
+  const { category: { id, name, postsCnt } } = props;
   const { deleteCategory } = useDeleteCategoryMutation();
   const { updateCategory } = useUpdateCategoryMutation();
   const { formState: { errors }, register, handleSubmit, setError, clearErrors } = useForm<CategoryInputs>({ defaultValues: { name: name } });
 
+  const [show, setShow] = useState(false);
+  const handleClose = () => setShow(false);
+  const handleShow = () => setShow(true);
+
   const isErrorCategoryAttributes = (attribute: string): attribute is CategoryAttributes => attribute.includes(attribute);
-  const setValidationErrors = (errors, errorFullMessages: string[]) => {
+  const setValidationErrors = (errors) => {
     clearErrors();
 
-    setFullMessages(errorFullMessages);
     errors.forEach(error => {
       const { attribute } = error;
       if (isErrorCategoryAttributes(attribute)) setError(attribute, { message: error.messages.join(' ') })
@@ -42,12 +48,12 @@ export default function Category(props: Props) {
   }
 
   const onSubmit = async (input: any) => {
+    handleClose()
     try {
       const res = await updateCategory({ variables: { input: {id, ...input} }});
-      console.log(res)
-      const { errors, fullMessages } = res.data.updateCategory
+      const { errors } = res.data.updateCategory
       if (errors && errors.length !== 0) {
-        setValidationErrors(errors, fullMessages);
+        setValidationErrors(errors);
       } else {
         clearErrors();
       }
@@ -59,29 +65,43 @@ export default function Category(props: Props) {
   }
 
   const callDeleteCategory = (id: string) => {
-    try {
-      deleteCategory({ variables: { input: { id } } })
-    }
-    catch(error){
-      // alert(`システムエラーが発生しました。\n${error}`)
-      console.log(error)
+    if(window.confirm('カテゴリ内の投稿も削除されますがよろしいですか？')){
+      try {
+        deleteCategory({ variables: { input: { id } } })
+      }
+      catch(error){
+        // alert(`システムエラーが発生しました。\n${error}`)
+        console.log(error)
+      }
     }
   };
 
   return (
-    <div>
-      <FullMessages fullMessages={fullMessages}/>
+    <li className="list-group-item d-flex justify-content-between align-items-center">
       <Link to={`/categories/${id}`}>{name}</Link>
+      <Badge pill bg="secondary">{postsCnt}</Badge>
       {
         currentAdmin &&
         <>
-          <button onClick={() => callDeleteCategory(id)}>削除する</button>
-          <form onSubmit={handleSubmit(onSubmit)}>
-            <TextField id='name' label='name' register={register('name')} errorText={errors.name?.message} />
-            <SubmitButton label='編集する'/>
-          </form>
+          <Button variant="danger" onClick={() => callDeleteCategory(id)}>削除する</Button>
+          <Button variant="primary" onClick={handleShow}>編集する</Button>
+          {/* Modal */}
+          <Modal show={show} onHide={handleClose}>
+            <form onSubmit={handleSubmit(onSubmit)}>
+              <Modal.Header closeButton>
+                <Modal.Title>カテゴリ編集</Modal.Title>
+              </Modal.Header>
+              <Modal.Body>
+                <TextField klass='' id='name' label='カテゴリ名' register={register('name')} errorText={errors.name?.message} />
+              </Modal.Body>
+              <Modal.Footer>
+                <Button variant="secondary" onClick={handleClose}>閉じる</Button>
+                <SubmitButton label='更新'/>
+              </Modal.Footer>
+            </form>
+          </Modal>
         </>
       }
-    </div>
+    </li>
   )
 }
